starlark: require go1.23 for new iterators
The correct type for new iterators is iter.Seq, which
depends on a go1.23-only package, so we must limit the
new functions to new toolchains.

In principle this is (yet another) breaking change,
but I suspect only users of (pre-)go1.23 will notice,
and they are living on the bleeding edge by choice.

Also, fix a couple of doc links.
